+++ components/clinical/shared/summary/src/organisms/missing-data-summary.tsx
@@ -7,8 +7,12 @@ const MissingDataSummary: FC<Props> = ({ type, noData, partialData, completeData
   const missingDataSection =
     missingData != null && missingData.length > 0 ? (
       <>
-        <Card.Banner data-testid="missing-data-count" status="warning">
-          <CounterIcon size="large" status="amber" value={missingData.length} /> Missing Data
+        <Card.Banner
+          data-testid="missing-data-count"
+          status="warning"
+          icon={<CounterIcon size="large" status="amber" value={missingData.length} />}
+        >
+          Missing Data
         </Card.Banner>
         <Card.List data-testid="missing-data-list">
           {missingData.map((data, index) => (
@@ -23,8 +27,12 @@ const MissingDataSummary: FC<Props> = ({ type, noData, partialData, completeData
   const partialDataSection =
     partialData != null && partialData.length > 0 ? (
       <>
-        <Card.Banner data-testid="partial-data-count" status="warning">
-          <CounterIcon size="large" status="amber" value={partialData.length} /> Partial Data
+        <Card.Banner
+          data-testid="partial-data-count"
+          status="warning"
+          icon={<CounterIcon size="large" status="amber" value={partialData.length} />}
+        >
+          Partial Data
         </Card.Banner>
         <Card.List data-testid="partial-data-list">
           {partialData.map((data, index) => (
@@ -39,8 +47,12 @@ const MissingDataSummary: FC<Props> = ({ type, noData, partialData, completeData
   const completeDataSection =
     completeData != null && completeData.length > 0 ? (
       <>
-        <Card.Banner data-testid="complete-data-count" status="info">
-          <CounterIcon size="large" status="info" value={completeData.length} /> Complete Data
+        <Card.Banner
+          data-testid="complete-data-count"
+          status="info"
+          icon={<CounterIcon size="large" status="info" value={completeData.length} />}
+        >
+          Complete Data
         </Card.Banner>
         <Card.List data-testid="complete-data-list">
           {completeData.map((data, index) => (
@@ -55,8 +67,12 @@ const MissingDataSummary: FC<Props> = ({ type, noData, partialData, completeData
   const noDataSection =
     noData != null && noData.length > 0 ? (
       <>
-        <Card.Banner data-testid="no-data-count" status="info">
-          <CounterIcon size="large" status="info" value={noData.length} /> No Data
+        <Card.Banner
+          data-testid="no-data-count"
+          status="info"
+          icon={<CounterIcon size="large" status="info" value={noData.length} />}
+        >
+          No Data
         </Card.Banner>
         <Card.List data-testid="no-data-list">
           {noData.map((data, index) => (