+++ packages/types/src/atoms/local-types.ts
@@ -8,3 +8,9 @@ export enum PatientIdentifierType {
   NhsNumber = 'https://fhir.nhs.uk/Id/nhs-number',
   PasNumber = 'https://leedsth.nhs.uk/Id/pas-number',
 }
+
+export enum CodeSystem {
+  MedicationTypeIdentifier = 'https://fhir.leedsth.nhs.uk/CodeSystem/medication-type-identifier-1',
+}
+
+export type StatusTypes = 'info' | 'danger' | 'warning'